checkpoint: extract headers and query params from request
